+++ NEWS.md
@@ -1,5 +1,7 @@
 # vroom (development version)
 
+* vroom now offers to install the archive package if it's needed to complete the user's request (https://github.com/tidyverse/readr/issues/1334).
+
 * Unclosed quotes (e.g., `a,b,"c` with no closing `"`) now trigger a warning, instead of silent data truncation. The affected row is also newly included in the returned data, which should facilitate troubleshooting (#484, https://github.com/tidyverse/readr/issues/1539, https://github.com/tidyverse/readr/issues/1491).
 
 * [vroom.tidyverse.org](https://vroom.tidyverse.org/) is the new home of